requirements.txt
Pillow==10.0.1
ttkbootstrap==1.10.0
piexif==1.1.3
